Beta X.X

+++ Backend/Controllers/Controller.cs
@@ -1,5 +1,10 @@
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
 using Backend;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.Tokens;
+using Swashbuckle.AspNetCore.Annotations;
 
 namespace WebApplication1.Controllers;
 
@@ -9,27 +14,54 @@ namespace WebApplication1.Controllers;
 [Produces("application/json")]
 [ProducesResponseType(StatusCodes.Status200OK)]
 [ProducesResponseType(StatusCodes.Status400BadRequest)]
+[ProducesResponseType(StatusCodes.Status401Unauthorized)]
 public class Controller(ILogger<Controller> logger) : ControllerBase
 {
     private readonly UserRepository _userRepository = UserRepositoryFactory.CreateUserRepository();
     private readonly ILogger _logger = logger;
 
     /// <summary>
-    /// Регистрация пользователя
+    /// Регистрация пользователя и возврат JWT токена
+    /// Если пользователь уже зарегистрирован, то возвращает JWT токен
     /// </summary>
     /// <returns>Токен авторизации</returns>
     [HttpPost]
     public async Task<IActionResult> CreateUserAsync([FromBody] User user)
     {
         var path = Request.Path + " " + Request.Method;
-        var userId = await _userRepository.AddUserAsync(user);
+        var userExists = await _userRepository.GetUserAsync(user.Login);
         try
         {
-            _logger.LogInformation("{Path}: User {UserId} created successfully", path, user.Login);
-            return Ok(new
+            if (userExists == 0)
             {
-                Message = $"User created successfully with ID: {userId}",
-                Token = "token"
+                var userId = await _userRepository.AddUserAsync(user);
+                _logger.LogInformation("{Path}: User {UserId} created successfully", path, user.Login);
+                await _userRepository.AddUserHistoryAsync(new UserHistory(user.Login,
+                    "Регистрация", ""));
+                return Ok(new
+                {
+                    Message = $"User created successfully with ID: {userId}",
+                    UserId = userId,
+                    Token = CreateJwt(user.Login)
+                });
+            }
+
+            var userID = await _userRepository.CheckUserAsync(user);
+            if (userID != 0)
+            {
+                await _userRepository.AddUserHistoryAsync(new UserHistory(user.Login,
+                    "Вход", "Возвращен токен"));
+                _logger.LogInformation("{Path}: User successfully enter. Return the token", path);
+                return Ok(new
+                {
+                    UserId =  userID,
+                    Token = CreateJwt(user.Login)
+                });
+            }
+            _logger.LogInformation("{Path}: User enter incorrect login or password", path);
+            return Unauthorized(new
+            {
+                Message = "incorrect login or password"
             });
         }
         catch (Exception e)
@@ -46,9 +78,18 @@ public class Controller(ILogger<Controller> logger) : ControllerBase
     /// <summary>
     /// Получить историю пользователя
     /// </summary>
-    [HttpGet]
-    public async Task<IActionResult> GetUserHistoryAsync(string id)
+    [HttpGet("id")]
+    [Authorize]
+    public async Task<IActionResult> GetUserHistoryAsync(int id)
     {
+        var login = User.Identity?.Name;
+        if (login == null || await _userRepository.GetUserAsync(login) != id)
+        {
+            return Unauthorized(new
+            {
+                Message = "Вы не можете получить историю другого пользователя"
+            });
+        }
         var path = Request.Path + " " + Request.Method;
         var userHistory = await _userRepository.GetUserHistoryAsync(id);
         _logger.LogInformation("{Path}: User history retrieved", path);
@@ -68,12 +109,22 @@ public class Controller(ILogger<Controller> logger) : ControllerBase
                 Message = "User history is empty",
             });
     }
+
     /// <summary>
     /// Удалить историю пользователя
     /// </summary>
-    [HttpDelete]
-    public async Task<IActionResult> DeleteHistoryUserAsync(string id)
+    [HttpDelete("id")]
+    [Authorize]
+    public async Task<IActionResult> DeleteHistoryUserAsync(int id)
     {
+        var login = User.Identity?.Name;
+        if (login == null || await _userRepository.GetUserAsync(login) != id)
+        {
+            return Unauthorized(new
+            {
+                Message = "Вы не можете удалить историю другого пользователя"
+            });
+        }
         var path = Request.Path + " " + Request.Method;
         var affectedRows = await _userRepository.DeleteUserHistoryAsync(id);
         _logger.LogInformation("{Path}: User history deleted", path);
@@ -91,24 +142,26 @@ public class Controller(ILogger<Controller> logger) : ControllerBase
                 Message = "Deleted rows: " + affectedRows,
             });
     }
-    
     /// <summary>
     /// Изменить пароль
     /// </summary>
     [HttpPatch]
-    public async Task<IActionResult> PatchPasswordAsync([FromBody] UserChangePassword request)
+    public async Task<IActionResult> PatchPasswordAsync([FromBody] UserChangesPassword request)
     {
         var path = Request.Path + " " + Request.Method;
         try
         {
-            await _userRepository.ChangeUserPasswordAsync(request.NewPassword, request.Login);
+            await _userRepository.ChangeUserPasswordAsync(request);
             _logger.LogInformation("{Path}: User password changed", path);
-    
+            await _userRepository.AddUserHistoryAsync(new UserHistory(request.Login,
+                "Смена пароля", $"Старый пароль{request.Password}"));
+
             return Ok(new
             {
                 Status = 200,
                 Success = true,
                 Message = "User password changed",
+                Token = CreateJwt(request.Login)
             });
         }
         catch (Exception e)
@@ -122,4 +175,31 @@ public class Controller(ILogger<Controller> logger) : ControllerBase
         }
     }
 
+    [NonAction]
+    private string CreateJwt(string value)
+    {
+        var claims = new List<Claim> { new Claim(ClaimTypes.Name, value) };
+        var jwt = new JwtSecurityToken(
+            issuer: AuthOptions.ISSUER,
+            audience: AuthOptions.AUDIENCE,
+            claims: claims,
+            expires: DateTime.UtcNow.Add(TimeSpan.FromDays(1)),
+            signingCredentials: new SigningCredentials(AuthOptions.GetSymmetricSecurityKey(),
+                SecurityAlgorithms.HmacSha256));
+        var encodedJwt = new JwtSecurityTokenHandler().WriteToken(jwt);
+        return encodedJwt;
+    }
+
+    [NonAction]
+    private string? ReadJWT(HttpContext context)
+    {
+        var token = context.Request.Headers["Authorization"].FirstOrDefault();
+        if (string.IsNullOrEmpty(token))
+        {
+            return null; // токен не найден
+        }
+        var jwtToken = new JwtSecurityTokenHandler().ReadJwtToken(token);
+        var login = jwtToken.Claims.FirstOrDefault(claim => claim.Type == ClaimTypes.Name)?.Value;
+        return login;
+    }
 }
